Implement core event and broadcast classes for simulation.
Added detailed implementations for events (DetectObjectsEvent, PoseEvent, TrackedObjectsEvent) and broadcasts (CrashedBroadcast, TerminatedBroadcast, TickBroadcast). Updated CameraService and TimeService to handle these functionalities and support testing through additional constructors. Minor type adjustments (e.g., CloudPoint fields) for accuracy.

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -7,6 +7,6 @@ package bgu.spl.mics.application.objects;
 public class CloudPoint {
 
     // Fields
-    private int x; // X-coordinate
-    private int y; // Y-coordinate
+    private double x; // X-coordinate
+    private double y; // Y-coordinate
 }
